Facade for data manager in context_search
